requirements.txt
google-generativeai
pyttsx3
spotipy
speechrecognition
requests
flask
kivy
pypiwin32